+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.1.2 (2023-10-18)
+* Allow loading config from resolved class
+
 ## v0.1.1 (2023-10-18)
 * Refactored package file structure
 * Fixed Respository typo